Modified: architecture of application, Added lombok,Beans for creating queues
