+++ pom.xml
@@ -44,12 +44,15 @@
             <groupId>com.example</groupId>
             <artifactId>student-enrollment-multiconfig</artifactId>
             <version>1.0.0</version>
-            <scope>system</scope>
-            <systemPath>C:/Users/Ardak Shalkar/Development/UnitimeSchedulerWebApp/student-enrollment-multiconfig-1.0.0.jar</systemPath>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>annotationProcessor</scope>
         </dependency>
     </dependencies>
     <build>
