+++ src/main/java/com/example/unitimescheduler/StatusPublisher.java
@@ -1,19 +1,19 @@
 package com.example.unitimescheduler;
 
+import com.example.unitimescheduler.models.StatusMessage;
+import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class StatusPublisher {
     private final RabbitTemplate rabbitTemplate;
-    private final String statusQueue;
 
-    public StatusPublisher(RabbitTemplate rabbitTemplate,
-                           @Value("${job.status-queue}") String statusQueue) {
-        this.rabbitTemplate = rabbitTemplate;
-        this.statusQueue = statusQueue;
-    }
+    @Value("${job.status-queue}")
+    private String statusQueue;
+
 
     public void sendStatus(StatusMessage statusMessage) {
         rabbitTemplate.convertAndSend(statusQueue, statusMessage);
