+++ src/main/java/com/example/unitimescheduler/config/RabbitConfig.java
@@ -1,5 +1,6 @@
-package com.example.unitimescheduler;
+package com.example.unitimescheduler.config;
 
+import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,5 +10,15 @@ public class RabbitConfig {
     @Bean
     public Jackson2JsonMessageConverter jackson2JsonMessageConverter() {
         return new Jackson2JsonMessageConverter();
+    }
+
+    @Bean
+    public Queue jobQueue(){
+        return  new Queue("jobQueue", true);
+    }
+
+    @Bean
+    public Queue jobStatusQueue() {
+        return new Queue("jobStatusQueue", true);
     }
 }
