+++ src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.rabbitmq.host=localhost
+spring.rabbitmq.port=5672
+spring.rabbitmq.username=guest
+spring.rabbitmq.password=guest
+
+spring.application.name=unitime-scheduler
+
+job.queue=jobQueue
+job.status-queue=jobStatusQueue
