--- src/main/resources/application.yml
@@ -1,12 +1,0 @@
-spring:
-  rabbitmq:
-    host: localhost
-    port: 5672
-    username: guest
-    password: guest
-  application:
-    name: unitime-scheduler
-
-job:
-  queue: jobQueue
-  status-queue: jobStatusQueue
